#1349 PR fixes
