+++ src/app/graphql.module.ts
@@ -2,8 +2,9 @@ import { APOLLO_OPTIONS, ApolloModule } from 'apollo-angular';
 import { HttpLink } from 'apollo-angular/http';
 import { NgModule } from '@angular/core';
 import { ApolloClientOptions, InMemoryCache } from '@apollo/client/core';
+import { environment } from "../environments/environment";
 
-const uri = 'https://apidev.alor.ru/hyperion';
+const uri = environment.apiUrl + '/hyperion';
 export function createApollo(httpLink: HttpLink): ApolloClientOptions<any> {
   return {
     link: httpLink.create({ uri }),
