+++ src/app/modules/bond-screener/models/bond-screener.model.ts
@@ -6,7 +6,7 @@ export interface BondScreenerResponse {
 }
 
 export interface BondEdge {
-  node: BondNode;
+  node: Partial<BondNode>;
   cursor: string;
 }
