+++ src/app/shared/services/graph-ql.service.spec.ts
@@ -1,24 +1,29 @@
 import { TestBed } from '@angular/core/testing';
 
-import { BondScreenerService } from './bond-screener.service';
+import { GraphQlService } from './graph-ql.service';
 import { of } from "rxjs";
-import { GraphQlService } from "../../../shared/services/graph-ql.service";
+import { ErrorHandlerService } from "./handle-error/error-handler.service";
+import { Apollo } from "apollo-angular";
 
-describe('BondScreenerService', () => {
-  let service: BondScreenerService;
+describe('GraphQlService', () => {
+  let service: GraphQlService;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
       providers: [
         {
-          provide: GraphQlService,
+          provide: Apollo,
           useValue: {
             watchQuery: jasmine.createSpy('watchQuery').and.returnValue({ valueChanges: of({})})
           }
+        },
+        {
+          provide: ErrorHandlerService,
+          useValue: {}
         }
       ]
     });
-    service = TestBed.inject(BondScreenerService);
+    service = TestBed.inject(GraphQlService);
   });
 
   it('should be created', () => {
