--- src/assets/i18n/bond-screener/settings/en.json
@@ -1,5 +1,0 @@
-{
-  "selectColumnsError": "Select columns",
-  "selectColumnLabel": "Table columns",
-  "selectColumnPlaceholder": "Select..."
-}
