--- src/assets/i18n/bond-screener/settings/ru.json
@@ -1,6 +1,0 @@
-{
-  "selectColumnsError": "Выберите колонки",
-  "selectColumnLabel": "Колонки таблицы",
-  "selectColumnPlaceholder": "Выберите..."
-
-}
