+++ src/assets/i18n/en.json
@@ -75,12 +75,16 @@
     "instrumentGroup": {
       "label": "Trading mode",
       "commonError": "Select trading mode"
+    },
+    "tableColumns": {
+      "selectColumnLabel": "Table columns"
     }
   },
 
   "validationErrors": {
     "emptyFieldError": "Enter value",
     "selectValueError": "Select value",
+    "selectColumnsError": "Выберите колонки",
     "tooLittle": "Too little",
     "tooMuch": "Too much",
     "intervalError": "Value should be from {{min}} to {{max}}"
